Generate a fragment distance vector for your digested DNAString object

Generate a fragment distance vector for your digested DNAString object

Usage

FROG_nonpal(my_DNA, pattern1 = "ATGCTGATT", pattern2 = "CGTGACTGGA")

Arguments

my_DNA,

That's your DNAString you aim to digest

pattern1,

The non-PALINDROMIC sequence pattern the restriction enzyme 1 recognizes

pattern2,

OPTIONAL: The non-PALINDROMIC nucleotide sequence pattern the restriction enzyme 2 recognizes

Value

A vector containing fragment lengths

Examples

my_DNA<-gen_DNA(1000)
pattern1<-'ATGCTGATT'
pattern2<-'CGTGACTGGA'
FROG_pal(my_DNA, pattern1, pattern2)
#> [1] 999

library(BSgenome.Hsapiens.UCSC.hg38)
chr1_Hsapiens<-BSgenome.Hsapiens.UCSC.hg38$chr1

FROG_pal(chr1_Hsapiens, pattern1)
#>    [1]   393921   229892   221212    54620  1299822   223607   214061   331081
#>    [9]   894139   311973   342579   226949   111959   164245    49073   310941
#>   [17]    34052       37       37  1232218   129016    20509   284141    75784
#>   [25]   590758    30060   544156   276241   559607  1022149    22843    16126
#>   [33]   300966    73688  1142490   124693   289623   218806    22198    26348
#>   [41]    33758    62814   253811   457776   208880   437959   499275    23283
#>   [49]    97448    57479   121633    73120   147813     5650    87863   195362
#>   [57]    83928   765264    21747   132445   392453     6277     6305   568780
#>   [65]   160931      499   265952   205445   247810    81293   307761    49567
#>   [73]    78491   227456   120608    28076   170166    27948    30180    15119
#>   [81]    12364   192143    10497   105955   179343   213712   135770    83897
#>   [89]    16645   545862   212824    23299    64133     6581    87651    24331
#>   [97]    99761   295136   310449  1031053    78849   266506   532746    37507
#>  [105]   231605   365007   400273    23971   323473   264440   186910    53325
#>  [113]    49303    23944   150748   104904   245699   869525   478495   504552
#>  [121]    36727   599282   752239   172457   218526    31499    32321   265443
#>  [129]   194350   153064   700696   601918     6672   131498   129561    21438
#>  [137]   183864   208142   164684  1113243    25915   643192   175781   750035
#>  [145]    99793   290570   120829   135995   183466    81644    31335     8474
#>  [153]   286063   799316   458958    19359     2090    39737   259167   152564
#>  [161]   208839    82728   141490   256761   165557    32925   277171   243627
#>  [169]   292628    11799   258244     7293   142498    98326   151225    41118
#>  [177]    31443   128561   242831   117632   600031    48617    66594    97606
#>  [185]    60555     3376   161059   486048    62131    13198    34841    29561
#>  [193]   125683   109877   963849    59076     5718   221916    56291   106493
#>  [201]   727219   164802   307888   103341   156098   497707   139111    65068
#>  [209]   151525   215256   899754    49243    78503   202874   490117     5811
#>  [217]    11252     9094    52019   190730   187280    59326    47048    58779
#>  [225]   189230    82231   394823   247054   451652   978131    88546   224645
#>  [233]   227742   230380   291820   247779   341220   332735   112295   102874
#>  [241]    34206   205747   216941   724495    70538   264032   257289   103030
#>  [249]    95696    67726     1594     3031    48377   142870    33839    99426
#>  [257]   530982   185230    57157    58136    13074   297037    70370    21128
#>  [265]   386712   238506    71391   278248   109460    95595    36097    53626
#>  [273]   106892   160823    49706     4856     7311    32847    29770   495627
#>  [281]   102160    10943    90902   160030    32028   175370   175301   109520
#>  [289]   346743   164616    30695   148104    37556    18305   162320    61794
#>  [297]   372660    90876   290721    77372    73582   108020   139488   282127
#>  [305]    35309     7548    41155   467249   117152   139875    17036    86785
#>  [313]   112196    48802   326253    46824   112440    55461    37669     4448
#>  [321]   608988    54024   463203   283951    60171   708583    40359    72075
#>  [329]    91024    26845    70213   134483    40522    65203    73356   344567
#>  [337]    28675    44635   535266    89265    28149   111276    81374   271222
#>  [345]    49449   131948   140343    51128   160882   273679   104737   246575
#>  [353]    37444   292779   109277   218435    27498    12417    98239   184467
#>  [361]   125933    26808   280607   199306     8914   119414    78295    32991
#>  [369]     9001    57296   118633    77830    94245   167134   109258   245020
#>  [377]   608084    54829   208197    89790    12833   336696   213963   126350
#>  [385]    18723   562634    81233   187562   163166    34439   321773    39016
#>  [393]   191182   184978     5171   131372   170150    46155    78260    32680
#>  [401]    48604   384872    30806   367642    43185   177336   189422    18792
#>  [409]      313   153617   372621   196696    78604   571654    96406     2985
#>  [417]   189459   131877   120728     3347   184331    41727      327   109209
#>  [425]    12035   151131   211574   292757   380139    94978    26408     4587
#>  [433]     2689   137588    28270    23317    69088   120917    97715    31996
#>  [441]     6234    20955   119403    14972    74242    73215    45977   228783
#>  [449]   154176    49267    30081   101425   105864    29551    84123   301393
#>  [457]   155026     2963    90126   389671     9967   265228   115307    16569
#>  [465]    96862   289028   228368    15340     4952   102712   284847   544402
#>  [473]   113752    45506     7516    30637   240986   210748    95182   179556
#>  [481]   679785    92501   181012   126807    93325   220723    90084   286179
#>  [489]    50137   154138   184450   133563    82429    44177   240164    69224
#>  [497]    25878   703587       21    26552   631942    13530     5429   102360
#>  [505]   348024    76540    44574   570641   476210   244512    24859   400574
#>  [513]   228859   536052    47928    11418    23948   236447   146739    15581
#>  [521]   155909     1740    31411    27855   179422    22740    13760   208083
#>  [529]   425515      115    17300    58013   188920    12438     4661    83383
#>  [537]    37961   344573    77212    95016   130574    24885   325942    22205
#>  [545]    16216    40166   223177   398102   136682   234500   222274     8869
#>  [553]   283842    24862   391819   103984   241280   226434   337598   219718
#>  [561]   119510    13603    32724    21183    33811    67413     9857   194435
#>  [569]   264761   193465   118389    37267    33535    23984    41448   836814
#>  [577]   215145   174284    28456    87667   207727   248832    30518    19025
#>  [585]    19502    16102   244129     5337    28232    16141     3068   353815
#>  [593]    29936     3005   116664   472525   176975    52938    50189   212528
#>  [601]    46571     6547   130522   363573     5253    69539    13414    69487
#>  [609]   183330    40197    63029    37212   491159   541235    83865   104641
#>  [617]   367454   132647    55272    47762   205040   356125   103929    37109
#>  [625]    19175   206006   108677   170191   341993    52153    27951   159004
#>  [633]   352865   338194   359923   313375    11589    37949    12276   124295
#>  [641]   124347   518848    18551   107852    15307   109012   292832    84567
#>  [649]    79966   262248   609809    44283   193033   364091    96590   522633
#>  [657]   349348   181562    30160   116140   163485    71349      655   492489
#>  [665]   287606    60252   108651   113791   149068   266798   173571    96859
#>  [673]   150788   154310    93066   227419   175174    48350   123835     9638
#>  [681]    93963   115729   112142    48997    35417   306440    93546    80016
#>  [689]     4848   197769    23464    68625    31291   298875   218453   161327
#>  [697]    71442   175066   374371   465283  3842778     4081     4957     3916
#>  [705] 18317684   127548   698081   379311   351039    85158    65644   400362
#>  [713]    75676   137582    65367   352674   341773    54810    30487   349772
#>  [721]    79412    44630   168955   108124    88103    66025    56988    40887
#>  [729]   162784   174497   298797   384695   162681    71871    67779   583935
#>  [737]   290489   104838   695960   192700    78618    84923   296477   448715
#>  [745]    49956   123744   247090    17272   372254   100025    12766    78778
#>  [753]    18988   567631   439292    87476   132087   327219    73025   315076
#>  [761]   418439    81990     6687    11863     2076     8299     5521   694620
#>  [769]   170300    93370   470717   114241   447858   259538   164620    78631
#>  [777]     5074    15299   102255    21109    36971   353360   209212   310400
#>  [785]    81261   184310   119120    32748     7574   103534    49654   155571
#>  [793]   306374   304040   115623    75160    92483   401196   170523   289568
#>  [801]   149990   267985      324   154604   408499   111763     7214   361563
#>  [809]   219724     6047    80285   173644   148198   118845   412408    23980
#>  [817]   141351   172434   458376   500919   108221   328196   341295    43663
#>  [825]   664997   100799    37881   341608   381616    56765    70698   211796
#>  [833]      328    61979   648611   381350    62456    65461    88178    42289
#>  [841]     1469    85133   111278   375370    47365    84999   225940    60492
#>  [849]   246720    17599    56960      784   126650    59209    32587   203668
#>  [857]    96473    30043    57309      469   157429   178503   259233    58324
#>  [865]    44331    63851    25092    56178   132014   121013    87427   319273
#>  [873]    27075   368528   209388   693417    97024   243892     2479   269805
#>  [881]    65203    88147      348   133721   274712   109517   151371    85499
#>  [889]   448818   296379   144522     9590    70669    94312    36913    49441
#>  [897]    95388   381079   300628   130163    34347   218281   262068   136958
#>  [905]   331179    41136    39578   329928     6169   148939   424212     6982
#>  [913]    21286   388499   565931    92416    79129   468473    38227    73955
#>  [921]    97100    38149    82416    11303     1214    17861    19672    30522
#>  [929]   228081   245101    14939   102402   208424    32720     7475     1528
#>  [937]    32407   122992   319568   129852   486355   208805   195856   192654
#>  [945]   448461   200695   169274     2959   406337    18097    52268   219717
#>  [953]    35538    87037    31990    45751   411829   204023   108590    81065
#>  [961]    22640    96585    35747   411485    68273    25641    30759   275081
#>  [969]   187158   131254    93245    36098    47430   190667   116282    79633
#>  [977]    21820    47844   123104    91213   166545   125010   523099    25292
#>  [985]   302719     1112    54432   137804    70719    14574    66853   360170
#>  [993]    91016   549181    17775   318297    40532     2265    56337   171604
#>  [ reached getOption("max.print") -- omitted 362 entries ]